lpr_engine/requirements.txt
absl-py
cachetools
certifi
charset-normalizer
colorama
cycler
easyocr
fonttools
google-auth
google-auth-oauthlib
grpcio
idna
imageio
importlib-metadata
kiwisolver
Markdown
matplotlib
networkx
numpy
oauthlib
opencv-contrib-python
opencv-python
packaging
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
pytesseract
python-bidi
python-dateutil
pytz
PyWavelets
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scipy
seaborn
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tifffile
torch
torchaudio
torchvision
tqdm
typing_extensions
urllib3
Werkzeug
wincertstore
zipp